Revert "Add the ability to record any key/value pair in the 'custom' field in metadata.yaml (#976)"
This reverts commit e156eb9e38c377df33e5075ab7fa48f3b52eaa12.
