+++ rosbag2_cpp/test/rosbag2_cpp/test_info.cpp
@@ -98,104 +98,6 @@ TEST_F(TemporaryDirectoryFixture, read_metadata_supports_version_2) {
   }
 }
 
-TEST_F(TemporaryDirectoryFixture, read_metadata_supports_version_6) {
-  const std::string bagfile = "rosbag2_bagfile_information:\n"
-    "  version: 6\n"
-    "  storage_identifier: sqlite3\n"
-    "  relative_file_paths:\n"
-    "    - some_relative_path\n"
-    "    - some_other_relative_path\n"
-    "  duration:\n"
-    "    nanoseconds: 100\n"
-    "  starting_time:\n"
-    "    nanoseconds_since_epoch: 1000000\n"
-    "  message_count: 50\n"
-    "  topics_with_message_count:\n"
-    "    - topic_metadata:\n"
-    "        name: topic1\n"
-    "        type: type1\n"
-    "        serialization_format: rmw1\n"
-    "        offered_qos_profiles: default\n"
-    "      message_count: 100\n"
-    "    - topic_metadata:\n"
-    "        name: topic2\n"
-    "        type: type2\n"
-    "        serialization_format: rmw2\n"
-    "        offered_qos_profiles: default\n"
-    "      message_count: 200\n"
-    "  compression_format: \"zstd\"\n"
-    "  compression_mode: \"FILE\"\n"
-    "  files:\n"
-    "    - path: test.db3\n"
-    "      starting_time:\n"
-    "        nanoseconds_since_epoch: 0\n"
-    "      duration:\n"
-    "        nanoseconds: 100000000\n"
-    "      message_count: 200\n"
-    "  custom_data:\n"
-    "    key1: value1\n"
-    "    key2: value2\n";
-
-  {
-    std::ofstream fout {
-      (rcpputils::fs::path(temporary_dir_path_) / rosbag2_storage::MetadataIo::metadata_filename)
-      .string()};
-    fout << bagfile;
-  }
-
-  rosbag2_cpp::Info info;
-  const auto metadata = info.read_metadata(temporary_dir_path_, "sqlite3");
-
-  EXPECT_EQ(metadata.version, 6);
-  EXPECT_EQ(metadata.storage_identifier, "sqlite3");
-
-  const auto expected_paths =
-    std::vector<std::string>{"some_relative_path", "some_other_relative_path"};
-  EXPECT_EQ(metadata.relative_file_paths, expected_paths);
-  EXPECT_EQ(metadata.duration, std::chrono::nanoseconds{100});
-  EXPECT_EQ(
-    metadata.starting_time,
-    std::chrono::time_point<std::chrono::high_resolution_clock>{std::chrono::nanoseconds{1000000}});
-  EXPECT_EQ(metadata.message_count, 50u);
-  EXPECT_EQ(metadata.topics_with_message_count.size(), 2u);
-
-  {
-    auto it1 = metadata.custom_data.find("key1");
-    ASSERT_NE(it1, metadata.custom_data.end());
-    EXPECT_STREQ(it1->second.c_str(), "value1");
-
-    auto it2 = metadata.custom_data.find("key2");
-    ASSERT_NE(it2, metadata.custom_data.end());
-    EXPECT_STREQ(it2->second.c_str(), "value2");
-  }
-
-  {
-    EXPECT_EQ(metadata.files.size(), 1U);
-    EXPECT_EQ(metadata.files[0].path, "test.db3");
-    EXPECT_EQ(metadata.files[0].starting_time.time_since_epoch().count(), 0U);
-    EXPECT_EQ(metadata.files[0].duration, std::chrono::nanoseconds{100000000});
-    EXPECT_EQ(metadata.files[0].message_count, 200U);
-  }
-
-  {
-    const auto actual_first_topic = metadata.topics_with_message_count[0];
-    const auto expected_first_topic =
-      rosbag2_storage::TopicInformation{{"topic1", "type1", "rmw1", ""}, 100};
-
-    EXPECT_EQ(actual_first_topic.topic_metadata, expected_first_topic.topic_metadata);
-    EXPECT_EQ(actual_first_topic.message_count, expected_first_topic.message_count);
-  }
-
-  {
-    const auto actual_second_topic = metadata.topics_with_message_count[1];
-    const auto expected_second_topic =
-      rosbag2_storage::TopicInformation{{"topic2", "type2", "rmw2", ""}, 200};
-
-    EXPECT_EQ(actual_second_topic.topic_metadata, expected_second_topic.topic_metadata);
-    EXPECT_EQ(actual_second_topic.message_count, expected_second_topic.message_count);
-  }
-}
-
 TEST_F(TemporaryDirectoryFixture, read_metadata_makes_appropriate_call_to_metadata_io_method) {
   std::string bagfile(
     "rosbag2_bagfile_information:\n"
