+++ rosbag2_py/src/rosbag2_py/_storage.cpp
@@ -40,12 +40,10 @@ PYBIND11_MODULE(_storage, m) {
     "output_serialization_format",
     &rosbag2_cpp::ConverterOptions::output_serialization_format);
 
-  using KEY_VALUE_MAP = std::unordered_map<std::string, std::string>;
   pybind11::class_<rosbag2_storage::StorageOptions>(m, "StorageOptions")
   .def(
     pybind11::init<
-      std::string, std::string, uint64_t, uint64_t, uint64_t, std::string, std::string, bool,
-      KEY_VALUE_MAP>(),
+      std::string, std::string, uint64_t, uint64_t, uint64_t, std::string, std::string, bool>(),
     pybind11::arg("uri"),
     pybind11::arg("storage_id") = "",
     pybind11::arg("max_bagfile_size") = 0,
@@ -53,8 +51,7 @@ PYBIND11_MODULE(_storage, m) {
     pybind11::arg("max_cache_size") = 0,
     pybind11::arg("storage_preset_profile") = "",
     pybind11::arg("storage_config_uri") = "",
-    pybind11::arg("snapshot_mode") = false,
-    pybind11::arg("custom_data") = KEY_VALUE_MAP{})
+    pybind11::arg("snapshot_mode") = false)
   .def_readwrite("uri", &rosbag2_storage::StorageOptions::uri)
   .def_readwrite("storage_id", &rosbag2_storage::StorageOptions::storage_id)
   .def_readwrite(
@@ -74,10 +71,7 @@ PYBIND11_MODULE(_storage, m) {
     &rosbag2_storage::StorageOptions::storage_config_uri)
   .def_readwrite(
     "snapshot_mode",
-    &rosbag2_storage::StorageOptions::snapshot_mode)
-  .def_readwrite(
-    "custom_data",
-    &rosbag2_storage::StorageOptions::custom_data);
+    &rosbag2_storage::StorageOptions::snapshot_mode);
 
   pybind11::class_<rosbag2_storage::StorageFilter>(m, "StorageFilter")
   .def(
