+++ rosbag2_storage/src/rosbag2_storage/storage_options.cpp
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #include <string>
-#include <unordered_map>
 
 #include "rosbag2_storage/storage_options.hpp"
 
@@ -32,7 +31,6 @@ Node convert<rosbag2_storage::StorageOptions>::encode(
   node["storage_preset_profile"] = storage_options.storage_preset_profile;
   node["storage_config_uri"] = storage_options.storage_config_uri;
   node["snapshot_mode"] = storage_options.snapshot_mode;
-  node["custom_data"] = storage_options.custom_data;
   return node;
 }
 
@@ -48,8 +46,6 @@ bool convert<rosbag2_storage::StorageOptions>::decode(
     node, "storage_preset_profile", storage_options.storage_preset_profile);
   optional_assign<std::string>(node, "storage_config_uri", storage_options.storage_config_uri);
   optional_assign<bool>(node, "snapshot_mode", storage_options.snapshot_mode);
-  using KEY_VALUE_MAP = std::unordered_map<std::string, std::string>;
-  optional_assign<KEY_VALUE_MAP>(node, "custom_data", storage_options.custom_data);
   return true;
 }
