+++ rosbag2_storage/test/rosbag2_storage/test_storage_options.cpp
@@ -29,8 +29,6 @@ TEST(storage_options, test_yaml_serialization)
   original.storage_preset_profile = "profile";
   original.storage_config_uri = "config_uri";
   original.snapshot_mode = true;
-  original.custom_data["key1"] = "value1";
-  original.custom_data["key2"] = "value2";
 
   auto node = YAML::convert<rosbag2_storage::StorageOptions>().encode(original);
 
@@ -48,5 +46,4 @@ TEST(storage_options, test_yaml_serialization)
   ASSERT_EQ(original.storage_preset_profile, reconstructed.storage_preset_profile);
   ASSERT_EQ(original.storage_config_uri, reconstructed.storage_config_uri);
   ASSERT_EQ(original.snapshot_mode, reconstructed.snapshot_mode);
-  ASSERT_EQ(original.custom_data, reconstructed.custom_data);
 }
