+++ rosbag2_tests/resources/reindex_test_bags/target_metadata/multiple_files/metadata.yaml
@@ -1,5 +1,5 @@
 rosbag2_bagfile_information:
-  version: 6
+  version: 5
   storage_identifier: sqlite3
   relative_file_paths:
     - multiple_files_0.db3
@@ -38,4 +38,3 @@ rosbag2_bagfile_information:
       duration:
         nanoseconds: 5499907705
       message_count: 12
-  custom_data: ~
